Fix crd too long issue
Since auto generated crd embed PodTemplateSpec, it results in some failures like `The CustomResourceDefinition "..." is invalid: metadata.annotations: Too long: must have at most 262144 characters`

Everytime we run `kubectl apply`, it will sets the kubectl.kubernetes.io/last-applied-configuration: '{...}' annotation on each object. The annotation contains the contents of the object configuration file that was used to create the object."

Using kubectl create helps workaround the issue but create will fail if crd exist so it need us to change logics to check and create.

In end end, we specify `maxDescLen` on controller-tools when generating the CRD and it helps reduce size.