Fix flaky tests by retrying 409 conflict error
Failure is because of error "Operation cannot be fulfilled on rayclusters.ray.io: the object has been modified; please apply your changes to the latest version and try again". This is pretty common and because test code may not use latest revision to update the object since operator updates the object in the middle. Retry on this kind of issues will work.