fix: use head group's ServiceAccount in autoscaler RoleBinding
instead of always using RayCluster name. Default to RayCluster
name if the ServiceAccount doesn't exist.

This is a follow up fix after
https://github.com/ray-project/kuberay/pull/259 and
https://github.com/ray-project/kuberay/pull/276.

The use case is when using the RayCluster autoscaler. The autoscaler
will use the head group's K8s ServiceAccount to make K8s API requests.
Always using the RayCluster name as the RoleBinding's subject name won't
be correct and will result in the autoscaler sidecar container not being
authorized by the K8s API.