refactor: log more descriptive info from initContainer
Right now the example workers' initContainer that checks for
head Pod to be running logs the message `waiting for myservice`.
There is no `myservice` per se. The actual thing it's waiting for
is the K8s Service for the head group.

This change updates the log message to be more descriptive and
removes the string `myservice` from various places to not be misleading.