Update ray-operator documentation and image version in ray-cluster.heterogeneous.yaml (#585)
The example program doesn't get executed as specified in the readme. This PR updated the readme to give users instructions to execute the example program and show what the result should look like. Besides, this PR upgrades the image to 2.0.0 to allow submitting jobs to Ray cluster, and removes object-manager-port and node-manager-port configuration to avoid port conflicts.

Signed-off-by: jasoonn <35865970+jasoonn@users.noreply.github.com>