optimize e2e test (#609)
Currently, we use time.sleep() to wait for the system convergence in E2E tests. However, time.sleep() will block the process for fixed units of time. Hence, we need to use new wait functions that block the process only when the system is not ready.

kind create cluster --wait ...
kubectl wait