[Feature][Doc] Explain that RBAC should be synchronized manually (#641)
Currently, RBAC YAML files in helm-chart/kuberay-operator/templates and ray-operator/config/rbac should be synchronized manually. We have pointed this out in DEVELOPMENT.md, but it is not easy for users to find the messages. Hence, we need to make this information become more explicit in both DEVELOPMENT.md and rbac-check.py.