fix: skip reconciling RayClusters whose metadata.generation field doesn't change
by adding `predicate.GenerationChangedPredicate` to reconciler.

> This predicate will skip update events that have no change in the object's
> metadata.generation field. The metadata.generation field of an object is
> incremented by the API server when writes are made to the spec field of an
> object. This allows a controller to ignore update events where the spec is
> unchanged, and only the metadata and/or status fields are changed.

https://pkg.go.dev/sigs.k8s.io/controller-runtime/pkg/predicate?utm_source=godoc#GenerationChangedPredicate

Without this change, the controller gets stuck in a tight loop repeatedly
updating the same RayCluster when a ResourceQuota is exhausted. The controller
updates the RayCluster.status, gets an event about the forbidden Pod with a
different name, updates the .status, gets an event, on and on.