[Feature] [RayJobs] Use finalizers to implement stopping a job upon cluster deletion (#735)
See #629 to get more context. The behavior of this PR is almost the same as #629. The only difference is that this PR promises that operator will try to stop the job at least once. In #629, if the RayJob is deleted when the operator is down, the operator will not try to stop the job.