[RayJob] Add runtime env YAML field (#1338)
Adds the field RuntimeEnvYAML to the RayJob CRD which accepts a multi-line YAML string. This format is preferred for two reasons:

Consistency with the ServeConfigV2 format, which is also a Ray configuration specified as a multi-line YAML string
(Related to above) Allows using snake_case fields without modification
We preserve the older field RuntimeEnv which accepts a base64-encoded string of the runtime env. We mark it as deprecated in the documentation.

We raise an error if both fields are specified.

Related issue number
Closes #1195

---------

Signed-off-by: Archit Kulkarni <architkulkarni@users.noreply.github.com>