[CI] Update latest ray version 2.5.0 -> 2.6.3 (#1320)
Update all files to Ray version 2.6.3 as in [2.5.0 Release] Change version numbers 2.4.0 -> 2.5.0 #1151
Update the Ray version used in the sample Job yaml test to Ray 2.6.2 and remove old ray versions that we don't guarantee compatibility with. This reduces costs and reduces the impact of flaky tests. We support the last three Ray stable versions and the last 1.x Ray version.

---------

Signed-off-by: Archit Kulkarni <architkulkarni@users.noreply.github.com>