[RayJob] Add default CPU and memory for job submitter pod (#1319)
Adds a default CPU and memory resource requirement for the pod that submits the RayJob.

image
Here's the memory and CPU usage of the job submitter pod with a job entrypoint script that prints "hello world" in a tight loop from Python. I tested it on a local kind cluster. Since the job submitter pod just calls ray job submit which runs the job remotely and streams the job output to stdout on the submitter pod, this kind of log streaming job should be the most stressful from the perspective of the job submitter pod.

Based on this, we set the requests to 500m CPU and 200MiB memory, and set the limits to 1 CPU and 1GiB memory.

Why are these changes needed?
It's good to provide reasonable defaults to reduce the burden on the user.

Related issue number
Closes #1198

---------

Signed-off-by: Archit Kulkarni <architkulkarni@users.noreply.github.com>