[fix][kubectl-plugin] error when creating cluster that already exists
Right now `kubectl ray create cluster NAME` returns 0 if a cluster with the
same name already exists. We want this command to have the same behavior as
`kubectl create` not `kubectl apply`. `kubectl create` exits with an error if
there's an existing resource with the same name.

Note: this change will break the existing behavior of users who run the command
as a way to update an existing cluster. Perhaps this is a desirable breaking
change.

## Before

```console
$ kubectl ray create cluster dxia-test1
Error: the Ray cluster dxia-test1 in namespace hyperkube already exists

$ echo $?
1
```

## After

```console
$ kubectl ray create cluster dxia-test1
Error: the Ray cluster dxia-test1 in namespace hyperkube already exists

$ echo $?
1
```

Signed-off-by: David Xia <david@davidxia.com>