Only try once in HTTP health check commands (#3469)
The BaseWgetHealthCommand is used to configure readiness and liveness
probes that Kubernetes will execute against the ray-worker container. These
checks should return fairly quickly so that _Kubernetes_ can get the
result and possibly retry them. By default, wget retries _20 times_. The
retrying should be left to Kubernetes, not wget, so that the command
doesn't just hang for minutes while Kubernetes waits for the result.